实装 shell 工具

+++ app/src/main/java/com/sevtinge/hyperceiler/ui/base/BaseSettingsActivity.java
@@ -1,21 +1,21 @@
 /*
-  * This file is part of HyperCeiler.
+ * This file is part of HyperCeiler.
 
-  * HyperCeiler is free software: you can redistribute it and/or modify
-  * it under the terms of the GNU Affero General Public License as
-  * published by the Free Software Foundation, either version 3 of the
-  * License.
+ * HyperCeiler is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License.
 
-  * This program is distributed in the hope that it will be useful,
-  * but WITHOUT ANY WARRANTY; without even the implied warranty of
-  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  * GNU Affero General Public License for more details.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
 
-  * You should have received a copy of the GNU Affero General Public License
-  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-  * Copyright (C) 2023-2024 HyperCeiler Contributions
-*/
+ * Copyright (C) 2023-2024 HyperCeiler Contributions
+ */
 package com.sevtinge.hyperceiler.ui.base;
 
 import android.content.Intent;
@@ -25,8 +25,9 @@ import android.text.TextUtils;
 import androidx.fragment.app.Fragment;
 
 import com.sevtinge.hyperceiler.R;
+import com.sevtinge.hyperceiler.callback.ITAG;
 import com.sevtinge.hyperceiler.utils.log.AndroidLogUtils;
-import com.sevtinge.hyperceiler.utils.shell.ShellUtils;
+import com.sevtinge.hyperceiler.utils.shell.ShellInit;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -101,38 +102,50 @@ public abstract class BaseSettingsActivity extends BaseActivity {
     public void doRestart(String[] packageName, boolean isRestartSystem) {
         boolean result = false;
         boolean pid = true;
-        if (isRestartSystem) {
-            result = ShellUtils.getResultBoolean("reboot", true);
-        } else {
-            if (packageName != null) {
-                for (String packageGet : packageName) {
-                    if (packageGet == null) {
-                        continue;
-                    }
-                    // String test = "XX";
-                    // ShellUtils.CommandResult commandResult = ShellUtils.execCommand("{ [[ $(pgrep -f '" + packageGet +
-                    //     "' | grep -v $$) != \"\" ]] && { pkill -l 9 -f \"" + packageGet +
-                    //     "\"; }; } || { echo \"kill error\"; }", true, true);
-                    ShellUtils.CommandResult commandResult =
-                        ShellUtils.execCommand("{ pid=$(pgrep -f '" + packageGet + "' | grep -v $$);" +
+        if (ShellInit.getShell() != null) {
+            if (isRestartSystem) {
+                result = ShellInit.getShell().append("reboot").sync().isResult();
+            } else {
+                if (packageName != null) {
+                    for (String packageGet : packageName) {
+                        if (packageGet == null) {
+                            continue;
+                        }
+                        // String test = "XX";
+                        // ShellUtils.CommandResult commandResult = ShellUtils.execCommand("{ [[ $(pgrep -f '" + packageGet +
+                        //     "' | grep -v $$) != \"\" ]] && { pkill -l 9 -f \"" + packageGet +
+                        //     "\"; }; } || { echo \"kill error\"; }", true, true);
+
+                        boolean getResult =
+                            ShellInit.getShell().append("{ pid=$(pgrep -f '" + packageGet + "' | grep -v $$);" +
                                 " [[ $pid != \"\" ]] && { pkill -l 9 -f \"" + packageGet + "\";" +
                                 " { [[ $? != 0 ]] && { killall -s 9 \"" + packageGet + "\" &>/dev/null;};}" +
                                 " || { { for i in $pid; do kill -s 9 \"$i\" &>/dev/null;done;};}" +
                                 " || { echo \"kill error\";};};}" +
-                                " || { echo \"kill error\";}",
-                            true, true);
-                    if (commandResult.result == 0) {
-                        if (commandResult.successMsg.equals("kill error")) {
-                            pid = false;
-                        } else result = true;
-                    } else
-                        AndroidLogUtils.LogE("doRestart: ", "result: " + commandResult.result +
-                            " errorMsg: " + commandResult.errorMsg + " package: " + packageGet, null);
+                                " || { echo \"kill error\";}").sync().isResult();
+                        ArrayList<String> outPut = ShellInit.getShell().getOutPut();
+                        ArrayList<String> error = ShellInit.getShell().getError();
+
+                        if (getResult) {
+                            if (outPut.size() != 0) {
+                                if (outPut.get(0).equals("kill error")) {
+                                    pid = false;
+                                } else {
+                                    result = true;
+                                }
+                            } else result = true;
+                        } else
+                            AndroidLogUtils.LogE("doRestart: ", "result: " + ShellInit.getShell().getResult() +
+                                " errorMsg: " + error + " package: " + packageGet, null);
+
+                    }
+                } else {
+                    AndroidLogUtils.LogE("doRestart: ", "packageName is null", null);
                 }
-            } else {
-                AndroidLogUtils.LogE("doRestart: ", "packageName is null", null);
+                // result = ShellUtils.getResultBoolean("pkill -l 9 -f " + packageName, true);
             }
-            // result = ShellUtils.getResultBoolean("pkill -l 9 -f " + packageName, true);
+        } else {
+            AndroidLogUtils.LogE(ITAG.TAG, "ShellExec is null!! from: doRestart");
         }
         if (!result) {
             new AlertDialog.Builder(this)
