+++ app/src/main/java/com/sevtinge/hyperceiler/view/RestartAlertDialog.java
@@ -1,21 +1,21 @@
 /*
-  * This file is part of HyperCeiler.
+ * This file is part of HyperCeiler.
 
-  * HyperCeiler is free software: you can redistribute it and/or modify
-  * it under the terms of the GNU Affero General Public License as
-  * published by the Free Software Foundation, either version 3 of the
-  * License.
+ * HyperCeiler is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License.
 
-  * This program is distributed in the hope that it will be useful,
-  * but WITHOUT ANY WARRANTY; without even the implied warranty of
-  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  * GNU Affero General Public License for more details.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
 
-  * You should have received a copy of the GNU Affero General Public License
-  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-  * Copyright (C) 2023-2024 HyperCeiler Contributions
-*/
+ * Copyright (C) 2023-2024 HyperCeiler Contributions
+ */
 package com.sevtinge.hyperceiler.view;
 
 import static com.sevtinge.hyperceiler.utils.api.VoyagerApisKt.isPad;
@@ -26,8 +26,10 @@ import android.content.Intent;
 import android.content.res.Resources;
 
 import com.sevtinge.hyperceiler.R;
+import com.sevtinge.hyperceiler.callback.ITAG;
 import com.sevtinge.hyperceiler.module.app.GlobalActions;
-import com.sevtinge.hyperceiler.utils.shell.ShellUtils;
+import com.sevtinge.hyperceiler.utils.log.AndroidLogUtils;
+import com.sevtinge.hyperceiler.utils.shell.ShellInit;
 
 import java.util.Arrays;
 import java.util.List;
@@ -59,13 +61,16 @@ public class RestartAlertDialog extends AlertDialog {
                     // ShellUtils.execCommand("pkill -l 9 -f " + mAppPackageNameList.get(i), true, false);
                     // String test = "XX";
                     String packageGet = mAppPackageNameList.get(i);
-                    ShellUtils.execCommand("{ pid=$(pgrep -f '" + packageGet + "' | grep -v $$);" +
+                    if (ShellInit.getShell() != null) {
+                        ShellInit.getShell().append("{ pid=$(pgrep -f '" + packageGet + "' | grep -v $$);" +
                             " [[ $pid != \"\" ]] && { pkill -l 9 -f \"" + packageGet + "\";" +
                             " { [[ $? != 0 ]] && { killall -s 9 \"" + packageGet + "\" &>/dev/null;};}" +
                             " || { { for i in $pid; do kill -s 9 \"$i\" &>/dev/null;done;};}" +
                             " || { echo \"kill error\";};};}" +
-                            " || { echo \"kill error\";}",
-                        true, false);
+                            " || { echo \"kill error\";}").sync();
+                    } else {
+                        AndroidLogUtils.LogE(ITAG.TAG, "ShellExec is null!! from: createMultipleChoiceView");
+                    }
                 }
             }
         });
